requirements.txt
torch>=2.4.0
safetensors>=0.4.3
numpy>=1.26.0
gradio>=6.0.0
pandas>=2.2.0
matplotlib>=3.9.0
seaborn>=0.13.0
psutil>=6.0.0
Pillow>=10.4.0